Send timers as gauges instead of absolute gauges.
The documentation on the difference between the two is minimal but it
seems that an absolute gauge will always overwrite the value with the
last one received. Based on this it would seem to be better to send a
gauge as Instrumental should store an average or combination of the
values.