Add a note to the readme about `flush_interval`.
We noticed odd issues with our production metrics and realized that
Instrumental was either averaging or overwriting metrics that were
flushed more often than the highest supported resolution (60s). To solve
this we increased the flush interval to allow the metrics to be
aggregated by statsite. This greatly improved the accuracy of our
metrics and should be noted for anyone using our sink.